[docker] move daemon to agent/docker.go
Beside install.go
